third_party/blink/web_tests/fast/events/popup-when-select-change-expected.txt

CONSOLE MESSAGE: PASSED
abcd
efgh
 If the pop-up was not blocked then there will be an PASS message. Otherwise, the test fails.
